requirements.txt
aenum==3.1.15
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.1.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
distro==1.9.0
Flask==3.0.0
frozenlist==1.4.1
future==0.18.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==3.7.0
MarkupSafe==2.1.3
multidict==6.0.4
openai==1.6.1
packaging==23.2
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
